Refactor OrderService
Heavily refactors what was the Orders class into an OrderService.  Where
previously the Orders class managed a single list of orders, requiring a
new Orders object for every new set of orders.  The OrderService will
now place and cancel any orders and only requires a single object to do
so.