Fix error when placing really small orders
Fixes an error that occurs when trying to place an order that will
return less than 0.01p.  This effected placement of any lay orders
below the normal Betfair minimum stake (the lower of £2 or 10 / price).

For a lay order of £0.50 @ 2:
The order service would first place a £2 lay at 1.01 and then try to
reduce the size of that order by £1.50, leaving £0.50 @ 1.01.
However £0.50 @ 1.1 returns less £0.01 and the reduction would fail.

Rather then using a default price of 1.01 to place lays under the minimum
stake, the lowest possible price for the size you have asked for is used.
In the case of placing £0.50 @ 2 the initial lay order is placed at
1.02.  As 1.02 will return £0.01 when it's size is reduced to £0.50.

The size of any lay order created is now also checked and adjusted if
needed so that any lay order placed will always return at least £0.01.
For example a lay order of £0.50 @ 1.01 will be automatically adjusted
to a size of £2 @ 1.01 so that it returns £0.01.