Fix Null report in LimitOrder
Fixes an issue in the LimitOrder class that would throw an exception
if the place report from Betfair is null for a given order.  This
happened when placing multiple orders and some of those order are below
the Betfair minimum.

When an order is below the Betfair minimum two orders are actually
placed.  When the second order is placed the OrderService class tries to
update all the orders, but only the orders that where below the minimum
recieve a place report as a result of this second order.