Fix invalid profit ratio error
Fixes the error when placing very small orders.

For lays, the absolute minimum profit allowed is £0.02.  And lay order
that needed to place an initial lay that resulted in a profit below that
amount was failing.  As a result any attempt to place a lay order that
returns less than £0.02 will not be placed as part of this fix.